+++ packages/orq-rc/src/models/operations/updatedataset.ts
@@ -182,7 +182,7 @@ export const UpdateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-03-25T09:16:36.980Z",
+    "2026-03-26T01:14:39.140Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {